Refactor tests to have channel tests in a mirroring file
Just next to channel.go.
