+++ rest/channel_test.go
@@ -2,31 +2,15 @@ package rest_test
 
 import (
 	"github.com/ably/ably-go/rest"
-	. "github.com/ably/ably-go/test/support"
 
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
-
-	_ "crypto/sha512"
 )
 
 var (
 	client  *rest.Client
 	channel *rest.Channel
 )
-
-var _ = BeforeSuite(func() {
-	_, err := TestAppInstance.Create()
-	Expect(err).NotTo(HaveOccurred())
-
-	client = rest.NewClient(TestAppInstance.Params)
-	channel = client.Channel("test")
-})
-
-var _ = AfterSuite(func() {
-	_, err := TestAppInstance.Delete()
-	Expect(err).NotTo(HaveOccurred())
-})
 
 var _ = Describe("Channel", func() {
 	var (
